# context/openapi_mybackend.json
{"openapi":"3.1.0","info":{"title":"FinPulse Experience API","version":"4.0.0"},"paths":{"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"Simple login endpoint. \nIn a real app, this would verify credentials.\nHere it just gets or creates a user.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/banks":{"get":{"tags":["banks"],"summary":"List Banks","operationId":"list_banks_api_banks_get","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/banks/{bank_id}/bootstrap":{"get":{"tags":["banks"],"summary":"Bootstrap Bank","operationId":"bootstrap_bank_api_banks__bank_id__bootstrap_get","parameters":[{"name":"bank_id","in":"path","required":true,"schema":{"type":"string","title":"Bank Id"}},{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/consent/initiate":{"post":{"tags":["consents"],"summary":"Initiate Consent","operationId":"initiate_consent_api_consent_initiate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentInitiateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/consent/initiate/product":{"post":{"tags":["consents"],"summary":"Initiate Product Consent","operationId":"initiate_product_consent_api_consent_initiate_product_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentInitiateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/consent/initiate/payment":{"post":{"tags":["consents"],"summary":"Initiate Payment Consent Endpoint","description":"Initiate payment consent for a bank.","operationId":"initiate_payment_consent_endpoint_api_consent_initiate_payment_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentInitiateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/consents/start":{"post":{"tags":["consents"],"summary":"Start Consent Alias","description":"Specification-friendly alias for /api/consent/initiate.","operationId":"start_consent_alias_api_consents_start_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentInitiateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/consent/callback":{"get":{"tags":["consents"],"summary":"Consent Callback","operationId":"consent_callback_api_consent_callback_get","parameters":[{"name":"consent_id","in":"query","required":true,"schema":{"type":"string","title":"Consent Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/consent/status":{"get":{"tags":["consents"],"summary":"Get Consent Status","operationId":"get_consent_status_api_consent_status_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"bank_id","in":"query","required":true,"schema":{"type":"string","title":"Bank Id"}},{"name":"request_id","in":"query","required":true,"schema":{"type":"string","title":"Request Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/consents/status":{"get":{"tags":["consents"],"summary":"Get Consent Status Alias","description":"Specification-friendly alias for /api/consent/status.","operationId":"get_consent_status_alias_api_consents_status_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"bank_id","in":"query","required":true,"schema":{"type":"string","title":"Bank Id"}},{"name":"request_id","in":"query","required":true,"schema":{"type":"string","title":"Request Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sync/start":{"post":{"tags":["sync"],"summary":"Start Sync Endpoint","description":"Запускает синхронизацию данных пользователя со всеми подключёнными банками.\n\nСинхронизация выполняется в фоне. Используйте GET /api/sync/status для проверки прогресса.\n\nReturns:\n    {\n        \"sync_id\": \"uuid...\",\n        \"status\": \"queued\"\n    }\n\nRaises:\n    409 Conflict: Если синхронизация уже запущена","operationId":"start_sync_endpoint_api_sync_start_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","description":"ID пользователя","title":"User Id"},"description":"ID пользователя"},{"name":"force","in":"query","required":false,"schema":{"type":"boolean","description":"Принудительное обновление (игнорировать кеш)","default":false,"title":"Force"},"description":"Принудительное обновление (игнорировать кеш)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sync/status":{"get":{"tags":["sync"],"summary":"Get Sync Status Endpoint","description":"Проверяет статус синхронизации пользователя.\n\nFrontend может опрашивать этот endpoint (polling) каждые 2-3 секунды.\n\nReturns:\n    Если синхронизация запущена:\n    {\n        \"status\": \"running\",\n        \"sync_id\": \"uuid...\",\n        \"locked_at\": \"2025-01-01T12:00:00\",\n        \"expires_at\": \"2025-01-01T12:05:00\"\n    }\n    \n    Если синхронизация завершена:\n    {\n        \"status\": \"completed\",\n        \"synced_at\": \"2025-01-01T12:00:00\"\n    }","operationId":"get_sync_status_endpoint_api_sync_status_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","description":"ID пользователя","title":"User Id"},"description":"ID пользователя"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/dashboard":{"get":{"tags":["analytics"],"summary":"Get Dashboard Metrics","description":"Получает dashboard метрики.\n\nQuery params:\n    user_id: ID пользователя\n    force_refresh: Принудительно обновить данные (игнорировать кеш)","operationId":"get_dashboard_metrics_api_dashboard_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"force_refresh","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Force Refresh"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/integration-status":{"get":{"tags":["analytics"],"summary":"Get Integration Status","operationId":"get_integration_status_api_integration_status_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntegrationStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/payments/mdp":{"post":{"tags":["payments"],"summary":"Pay Mdp","description":"Mandatory Daily Payment - обязательный платеж по кредиту.","operationId":"pay_mdp_api_payments_mdp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MDPPaymentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/payments/adp":{"post":{"tags":["payments"],"summary":"Pay Adp","description":"Additional Daily Payment - дополнительный платеж.","operationId":"pay_adp_api_payments_adp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ADPPaymentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/payments/sdp":{"post":{"tags":["payments"],"summary":"Pay Sdp","description":"Savings Daily Payment - пополнение накоплений.","operationId":"pay_sdp_api_payments_sdp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SDPPaymentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/onboarding/start":{"post":{"tags":["onboarding"],"summary":"Start Onboarding","description":"Начинает онбординг для пользователя.","operationId":"start_onboarding_api_onboarding_start_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardingStartRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/onboarding/status":{"get":{"tags":["onboarding"],"summary":"Get Onboarding Status","description":"Возвращает статус онбординга.","operationId":"get_onboarding_status_api_onboarding_status_get","parameters":[{"name":"onboarding_id","in":"query","required":true,"schema":{"type":"string","title":"Onboarding Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/onboarding/finalize":{"post":{"tags":["onboarding"],"summary":"Finalize Onboarding","description":"Завершает онбординг.","operationId":"finalize_onboarding_api_onboarding_finalize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardingFinalizeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/onboarding/consents":{"post":{"tags":["onboarding"],"summary":"Create All Consents","description":"Создает все необходимые consents для выбранных банков.","operationId":"create_all_consents_api_onboarding_consents_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardingConsentsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/onboarding/consents/status":{"get":{"tags":["onboarding"],"summary":"Get Consents Status","description":"Возвращает текущий статус всех согласий пользователя.","operationId":"get_consents_status_api_onboarding_consents_status_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/loans":{"get":{"tags":["loans"],"summary":"Get Loans","description":"Возвращает список всех кредитов пользователя с детальной информацией.\n\nВключает:\n- Список кредитов с полями: id, bank, type, balance, rate, monthly_payment, priority\n- Общие метрики: total_outstanding, mdp, adp\n- Стратегию ранжирования","operationId":"get_loans_api_loans_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","description":"User ID","title":"User Id"},"description":"User ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Loans Api Loans Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/deposits":{"get":{"tags":["loans"],"summary":"Get Deposits","description":"Возвращает список всех вкладов/депозитов пользователя.\n\nВключает:\n- Список депозитов с полями: id, bank, type, balance, rate, term_months\n- Общие метрики: total_saved, sdp, target, progress","operationId":"get_deposits_api_deposits_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","description":"User ID","title":"User Id"},"description":"User ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Deposits Api Deposits Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/refinance/optimize-loans":{"get":{"tags":["refinance"],"summary":"Optimize Loans","description":"Подбирает лучшие условия рефинансирования для кредитов пользователя.\n\nВозвращает:\n- Список предложений рефинансирования\n- Экономию в месяц\n- Рекомендации по стратегии","operationId":"optimize_loans_api_refinance_optimize_loans_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","description":"User ID","title":"User Id"},"description":"User ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Optimize Loans Api Refinance Optimize Loans Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/refinance/apply":{"post":{"tags":["refinance"],"summary":"Apply Refinance","description":"Подает заявку на рефинансирование.\n\nВ реальном приложении здесь был бы вызов банковского API для создания заявки.\nДля демо возвращаем успешный результат.","operationId":"apply_refinance_api_refinance_apply_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefinanceApplicationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Apply Refinance Api Refinance Apply Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ADPPaymentRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"bank_id":{"type":"string","title":"Bank Id"},"loan_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Loan Id"},"amount":{"type":"number","title":"Amount"}},"type":"object","required":["user_id","bank_id","amount"],"title":"ADPPaymentRequest"},"BankConsentRequest":{"properties":{"bank_id":{"type":"string","title":"Bank Id"},"consents":{"$ref":"#/components/schemas/BankConsents"}},"type":"object","required":["bank_id","consents"],"title":"BankConsentRequest"},"BankConsents":{"properties":{"account":{"type":"boolean","title":"Account"},"product":{"type":"boolean","title":"Product"},"payment":{"type":"boolean","title":"Payment","default":false}},"type":"object","required":["account","product"],"title":"BankConsents"},"BankPipelineStatus":{"properties":{"bank_id":{"type":"string","title":"Bank Id"},"bank_name":{"type":"string","title":"Bank Name"},"pipeline_status":{"type":"string","enum":["ok","partial","error"],"title":"Pipeline Status"},"steps":{"items":{"$ref":"#/components/schemas/PipelineStep"},"type":"array","title":"Steps"},"raw_metrics":{"$ref":"#/components/schemas/BankRawMetrics"}},"type":"object","required":["bank_id","bank_name","pipeline_status","steps","raw_metrics"],"title":"BankPipelineStatus"},"BankRawMetrics":{"properties":{"sum_account_balances":{"type":"number","title":"Sum Account Balances"},"sum_credit_debts":{"type":"number","title":"Sum Credit Debts"},"used_in_base_score":{"type":"boolean","title":"Used In Base Score"}},"type":"object","required":["sum_account_balances","sum_credit_debts","used_in_base_score"],"title":"BankRawMetrics"},"BaseScorePayload":{"properties":{"status":{"type":"string","enum":["ok","no_data","error"],"title":"Status"},"value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Value"},"currency":{"type":"string","title":"Currency"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["status","value","currency","reason"],"title":"BaseScorePayload"},"ConsentInitiateRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"bank_id":{"type":"string","title":"Bank Id"}},"type":"object","required":["user_id","bank_id"],"title":"ConsentInitiateRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IntegrationStatusResponse":{"properties":{"user_id":{"type":"string","title":"User Id"},"base_score":{"$ref":"#/components/schemas/BaseScorePayload"},"banks":{"items":{"$ref":"#/components/schemas/BankPipelineStatus"},"type":"array","title":"Banks"}},"type":"object","required":["user_id","base_score","banks"],"title":"IntegrationStatusResponse"},"LoginRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"user_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Name"}},"type":"object","required":["user_id"],"title":"LoginRequest"},"LoginResponse":{"properties":{"token":{"type":"string","title":"Token"},"profile":{"$ref":"#/components/schemas/UserProfile"}},"type":"object","required":["token","profile"],"title":"LoginResponse"},"MDPPaymentRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"bank_id":{"type":"string","title":"Bank Id"},"loan_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Loan Id"},"amount":{"type":"number","title":"Amount"}},"type":"object","required":["user_id","bank_id","amount"],"title":"MDPPaymentRequest"},"OnboardingConsentsRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"banks":{"items":{"$ref":"#/components/schemas/BankConsentRequest"},"type":"array","title":"Banks"}},"type":"object","required":["user_id","banks"],"title":"OnboardingConsentsRequest"},"OnboardingFinalizeRequest":{"properties":{"onboarding_id":{"type":"string","title":"Onboarding Id"},"user_id":{"type":"string","title":"User Id"}},"type":"object","required":["onboarding_id","user_id"],"title":"OnboardingFinalizeRequest"},"OnboardingStartRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"user_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Name"}},"type":"object","required":["user_id"],"title":"OnboardingStartRequest"},"PipelineStep":{"properties":{"name":{"type":"string","title":"Name"},"status":{"type":"string","enum":["ok","no_data","no_access","error"],"title":"Status"},"details":{"type":"string","title":"Details"},"error_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Code"}},"type":"object","required":["name","status","details"],"title":"PipelineStep"},"RefinanceApplicationRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"offer_id":{"type":"string","title":"Offer Id"},"loan_ids":{"items":{"type":"string"},"type":"array","title":"Loan Ids"},"phone":{"type":"string","title":"Phone"}},"type":"object","required":["user_id","offer_id","loan_ids","phone"],"title":"RefinanceApplicationRequest"},"SDPPaymentRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"bank_id":{"type":"string","title":"Bank Id"},"deposit_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Deposit Id"},"amount":{"type":"number","title":"Amount"}},"type":"object","required":["user_id","bank_id","amount"],"title":"SDPPaymentRequest"},"UserProfile":{"properties":{"user_id":{"type":"string","title":"User Id"},"name":{"type":"string","title":"Name"},"subscription_plan":{"type":"string","enum":["free","premium"],"title":"Subscription Plan","default":"free"},"is_active":{"type":"boolean","title":"Is Active","default":true}},"type":"object","required":["user_id","name"],"title":"UserProfile"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}